chore: reorganize lib
