+++ package.json
@@ -21,6 +21,7 @@
     "uuid": "^3.3.3"
   },
   "devDependencies": {
+    "@types/debug": "^4.1.5",
     "@types/jest": "^24.0.25",
     "@types/node": "^13.1.4",
     "@types/puppeteer": "^2.0.0",
