+++ src/lib/util.ts
@@ -1,14 +1,52 @@
-import { statPromise } from './misc';
+import fs from 'fs';
+import path from 'path';
+import util from 'util';
+import portfinder from 'portfinder';
+import debugConstructor from 'debug';
+
+export const debug = debugConstructor('vivliostyle-cli');
 
 export function log(...obj: any) {
   console.log('===>', ...obj);
 }
 
-export function statFile(filePath: string) {
-  return statPromise(filePath).catch((err) => {
+export async function statFile(filePath: string) {
+  try {
+    return util.promisify(fs.stat)(filePath);
+  } catch (err) {
     if (err.code === 'ENOENT') {
       throw new Error(`Specified input doesn't exists: ${filePath}`);
     }
     throw err;
+  }
+}
+
+export function findAvailablePort(): Promise<number> {
+  portfinder.basePort = 13000;
+  return portfinder.getPortPromise();
+}
+
+export function findEntryPointFile(
+  target: string,
+  root: string,
+): Promise<string> {
+  return new Promise((resolve) => {
+    const stat = fs.statSync(target);
+    if (!stat.isDirectory()) {
+      return resolve(path.relative(root, target));
+    }
+    const files = fs.readdirSync(target);
+    const index = [
+      'index.html',
+      'index.htm',
+      'index.xhtml',
+      'index.xht',
+    ].find((n) => files.includes(n));
+    if (index) {
+      return resolve(path.relative(root, path.resolve(target, index)));
+    }
+
+    // give up finding entrypoint
+    resolve(path.relative(root, target));
   });
 }
